requirements.txt
# Core dependencies for ms-swift fine-tuning
ms-swift
transformers>=4.51.0
datasets
accelerate
tensorboard
sentencepiece
protobuf
numpy<2.0.0
qwen-vl-utils

# Image processing for FIRE dataset preprocessing
pillow
tqdm

# Jupyter for interactive development
jupyter
jupyterlab
ipywidgets
